# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChatPRFAQ council",
    "version": "2.0.0"
  },
  "paths": {
    "/api/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_runs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/share/{token}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Shared Run",
        "operationId": "get_shared_run_api_share__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/framing/refine": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Refine Framing",
        "operationId": "refine_framing_api_runs__run_id__framing_refine_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineFramingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/framing/confirm": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Confirm Framing",
        "operationId": "confirm_framing_api_runs__run_id__framing_confirm_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmFramingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Cancel Run",
        "operationId": "cancel_run_api_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Resume Run",
        "operationId": "resume_run_api_runs__run_id__resume_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "brief",
              "market_research",
              "ledger",
              "problem_validation",
              "pr_v1",
              "pr_v2",
              "internal_faq",
              "concept_validation",
              "pr_v3",
              "external_faq",
              "prfaq",
              "bar_raiser",
              "plan"
            ],
            "title": "Kind"
          },
          "seat": {
            "type": "string",
            "enum": [
              "0",
              "1",
              "1b",
              "2",
              "3",
              "4",
              "5",
              "6",
              "7",
              "8",
              "9",
              "9b",
              "10"
            ],
            "title": "Seat"
          },
          "persona": {
            "type": "string",
            "enum": [
              "strategist",
              "analyst",
              "researcher",
              "pm",
              "vp_product",
              "biz_eng",
              "customer_success",
              "editor",
              "bar_raiser",
              "validation_lead"
            ],
            "title": "Persona"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "key_insight": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Insight"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "seat",
          "persona",
          "created_at",
          "key_insight",
          "payload"
        ],
        "title": "ArtifactRecord"
      },
...
      "ConfirmFramingRequest": {
        "properties": {
          "framing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Framing"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ConfirmFramingRequest"
      },
      "CreateRunRequest": {
        "properties": {
          "idea": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 10,
            "title": "Idea"
          }
        },
        "type": "object",
        "required": [
          "idea"
        ],
        "title": "CreateRunRequest"
      },
      "CreateRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "CreateRunResponse"
      },
...
      "Finding": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "claim": {
            "type": "string",
            "title": "Claim"
          },
          "metric": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric"
          },
          "topic": {
            "type": "string",
            "enum": [
              "market_size",
              "competition",
              "customer",
              "pricing",
              "trend",
              "regulation",
              "barrier",
              "success_factor",
              "other"
            ],
            "title": "Topic"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "claim",
          "metric",
          "topic",
          "source_ids",
          "confidence"
        ],
        "title": "Finding"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefineFramingRequest": {
        "properties": {
          "target_customer": {
            "type": "string",
            "title": "Target Customer",
            "default": ""
          },
          "customer_problem": {
            "type": "string",
            "title": "Customer Problem",
            "default": ""
          },
          "product_scope": {
            "type": "string",
            "title": "Product Scope",
            "default": ""
          }
        },
        "type": "object",
        "title": "RefineFramingRequest"
      },
...
      "RunListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "idea": {
            "type": "string",
            "title": "Idea"
          },
          "status": {
            "type": "string",
            "enum": [
              "framing",
              "awaiting_confirmation",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "working_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          }
        },
        "type": "object",
        "required": [
          "id",
          "idea",
          "status",
          "headline",
          "working_name",
          "created_at",
          "completed_at",
          "total_cost_usd"
        ],
        "title": "RunListItem"
      },
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunListItem"
            },
            "type": "array",
            "title": "Runs"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "runs",
          "next_cursor"
        ],
        "title": "RunListResponse"
      },
      "RunRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "idea": {
            "type": "string",
            "title": "Idea"
          },
          "status": {
            "type": "string",
            "enum": [
              "framing",
              "awaiting_confirmation",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "framing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Framing"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "share_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Token"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          }
        },
        "type": "object",
        "required": [
          "id",
          "idea",
          "status",
          "framing",
          "created_at",
          "updated_at",
          "completed_at",
          "total_cost_usd",
          "error",
          "config",
          "share_token",
          "headline"
        ],
        "title": "RunRecord"
      },
      "RunSnapshot": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunRecord"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepRecord"
            },
            "type": "array",
            "title": "Steps"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactRecord"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionRecord"
            },
            "type": "array",
            "title": "Versions"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "bar_raiser": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BarRaiserOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_seq": {
            "type": "integer",
            "title": "Last Seq"
          },
          "roster": {
            "items": {
              "$ref": "#/components/schemas/PersonaRecord"
            },
            "type": "array",
            "title": "Roster"
          },
          "seats": {
            "items": {
              "$ref": "#/components/schemas/SeatRecord"
            },
            "type": "array",
            "title": "Seats"
          }
        },
        "type": "object",
        "required": [
          "run",
          "steps",
          "artifacts",
          "versions",
          "findings",
          "sources",
          "bar_raiser",
          "last_seq",
          "roster",
          "seats"
        ],
        "title": "RunSnapshot"
      },
...
      "Source": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "cited_text": {
            "type": "string",
            "title": "Cited Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "cited_text"
        ],
        "title": "Source"
      },
...
      "StepRecord": {
        "properties": {
          "seat": {
            "type": "string",
            "enum": [
              "0",
              "1",
              "1b",
              "2",
              "3",
              "4",
              "5",
              "6",
              "7",
              "8",
              "9",
              "9b",
              "10"
            ],
            "title": "Seat"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "persona": {
            "type": "string",
            "enum": [
              "strategist",
              "analyst",
              "researcher",
              "pm",
              "vp_product",
              "biz_eng",
              "customer_success",
              "editor",
              "bar_raiser",
              "validation_lead"
            ],
            "title": "Persona"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "effort": {
            "type": "string",
            "title": "Effort"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd"
          },
          "fallback_used": {
            "type": "boolean",
            "title": "Fallback Used"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "key_insight": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Insight"
          }
        },
        "type": "object",
        "required": [
          "seat",
          "name",
          "persona",
          "attempt",
          "status",
          "model",
          "effort",
          "started_at",
          "ended_at",
          "duration_s",
          "usage",
          "cost_usd",
          "fallback_used",
          "error",
          "key_insight"
        ],
        "title": "StepRecord"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionRecord": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id"
          },
          "produced_by_seat": {
            "type": "string",
            "enum": [
              "0",
              "1",
              "1b",
              "2",
              "3",
              "4",
              "5",
              "6",
              "7",
              "8",
              "9",
              "9b",
              "10"
            ],
            "title": "Produced By Seat"
          },
          "persona": {
            "type": "string",
            "enum": [
              "strategist",
              "analyst",
              "researcher",
              "pm",
              "vp_product",
              "biz_eng",
              "customer_success",
              "editor",
              "bar_raiser",
              "validation_lead"
            ],
            "title": "Persona"
          },
          "press_release": {
            "$ref": "#/components/schemas/PressRelease"
          },
          "edits": {
            "items": {
              "$ref": "#/components/schemas/Edit"
            },
            "type": "array",
            "title": "Edits"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "version",
          "artifact_id",
          "produced_by_seat",
          "persona",
          "press_release",
          "edits",
          "caption"
        ],
        "title": "VersionRecord"
      }
    }
  }
}